requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
